refactor(siem): unify constants to services/siem-sources and export siemDeliveryReceipts
Hygiene cleanup after Waves 1–3 of the SIEM dual-read plan. Zero behavior
changes — dedupes constants to a single source of truth and fills a missing
schema export.

- services/siem-sources.ts is now the only definition of SIEM_SOURCES and
  CURSOR_INIT_SENTINEL
- Delete workers/siem-delivery-worker-constants.ts (cycle-break no longer
  needed once constants live in services/)
- siem-delivery-worker.ts: drop re-export shim and local SOURCES; import
  SIEM_SOURCES + CURSOR_INIT_SENTINEL from services/siem-sources
- siem-delivery-preflight.ts: drop private SOURCES duplicate; import
  SIEM_SOURCES from services/siem-sources
- siem-anchor-loader.ts + anchor-loader/delivery-worker tests: retarget
  CURSOR_INIT_SENTINEL import at services/siem-sources
- packages/db/src/index.ts: re-export siemDeliveryReceipts alongside
  siemDeliveryCursors so Drizzle consumers can import the table object

All 898 processor tests pass (including 430+ SIEM tests). Full monorepo
typecheck clean.
